[CAN-48] refactor: api 요청 수정
