+++ src/components/note/NoteModal.tsx
@@ -5,7 +5,8 @@ import { faFlag, faXmark } from '@fortawesome/free-solid-svg-icons';
 import { motion, AnimatePresence } from 'framer-motion';
 import ReactDOM from 'react-dom';
 import { useClickOutside } from '@/hooks/useClickOutside';
-import { useGoalTitle } from '@/hooks/useGoalTitle';
+import { useGoals } from '@/hooks/useGoals';
+import { Goal } from '@/types/goals';
 
 interface NoteModalProps {
   isOpen: boolean;
@@ -27,7 +28,11 @@ export default function NoteModal({
   goalId,
 }: NoteModalProps) {
   const [modalRef] = useClickOutside<HTMLDivElement>(onClose);
-  const { data: goalTitle } = useGoalTitle(goalId);
+  const { data: goals, isLoading } = useGoals();
+
+  const goalTitle = goals?.find(
+    (goal: Goal) => goal.goalId === Number(goalId),
+  )?.title;
 
   const modalContent = (
     <AnimatePresence>
@@ -63,7 +68,9 @@ export default function NoteModal({
                 <div className="flex size-8 items-center justify-center rounded-full bg-slate100 p-1 text-slate500">
                   <FontAwesomeIcon icon={faFlag} />
                 </div>
-                {goalTitle}
+                {isLoading
+                  ? '목표 로딩 중...'
+                  : goalTitle || '목표를 찾을 수 없습니다.'}
               </h1>
             </div>
